Add REST, and DB

+++ app/src/main/java/fr/lesmenusdumarche/lesmenusdumarche/domain/Seller.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.experimental.Builder;
+import lombok.Builder;
 
 /**
  * Created by maxime on 17/03/2016.
+ *
+ * Represents a seller
  */
 @Builder
 @NoArgsConstructor
@@ -16,7 +18,7 @@ public class Seller {
 
     @Getter
     @Setter
-    Double id;
+    Long id;
 
     @Getter
     @Setter
